Refactor rule editor directory structure
