+++ workspace/src/app/views/shared/RuleEditor/model/RuleEditorModel.utils.tsx
@@ -3,7 +3,7 @@ import { IHandleProps } from "gui-elements/src/extensions/react-flow/nodes/NodeD
 import { ArrowHeadType, Edge, FlowElement, Node, OnLoadParams, Position } from "react-flow-renderer";
 import { rangeArray } from "../../../../utils/basicUtils";
 import { IRuleNodeData, IRuleOperatorNode, NodeContentPropsWithBusinessData } from "../RuleEditor.typings";
-import { RuleNodeMenu } from "../ruleNode/RuleNodeMenu";
+import { RuleNodeMenu } from "../view/ruleNode/RuleNodeMenu";
 import { Tag, Highlighter, Spacing } from "gui-elements";
 import { RuleEditorNode } from "./RuleEditorModel.typings";
 import { Elements, XYPosition } from "react-flow-renderer/dist/types";
