+++ workspace/src/app/views/shared/RuleEditor/view/RuleEditorOperatorSidebar.tsx
