+++ workspace/src/app/views/shared/RuleEditor/view/RuleEditorView.typings.ts
