+++ workspace/src/app/views/shared/RuleEditor/view/ruleNode/RuleNodeParameters.tsx
@@ -1,0 +1,5 @@
+interface RuleNodeParametersProps {
+    update: {};
+}
+
+export const RuleNodeParameters = ({}: RuleNodeParametersProps) => {};
